arch/arm/src/imxrt/imxrt_usdhc.c: fixed no DMA build error

+++ arch/arm/src/imxrt/imxrt_usdhc.c
@@ -335,7 +335,7 @@ static void imxrt_blocksetup(FAR struct sdio_dev_s *dev,
 static int  imxrt_recvsetup(FAR struct sdio_dev_s *dev, FAR uint8_t *buffer,
               size_t nbytes);
 static int  imxrt_sendsetup(FAR struct sdio_dev_s *dev,
-              FAR const uint8_t *buffer, uint32_t nbytes);
+              FAR const uint8_t *buffer, size_t nbytes);
 #endif
 
 static int  imxrt_cancel(FAR struct sdio_dev_s *dev);
